+++ packages/cem-inheritance/src/index.ts
@@ -1,0 +1,3 @@
+export { updateCemInheritance } from "./cem-inheritance.js";
+export { cemInheritancePlugin } from "./cem-inheritance-plugin.js";
+export type { Options } from "./types.js";