+++ packages/jet-brains-integration/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 1.4.4
+
+- Allow any string for `descriptionSrc`
+
+## 1.4.3
+
+- Exclude private methods and properties using the `#` access modifier
+
+## 1.4.2
+
+- Deploy updated repo links
+
 ## 1.4.1
 
 - Update docs to reflect correct package name and imported module