+++ packages/jsx-integration/CHANGELOG.md
@@ -1,5 +1,21 @@
 # CHANGELOG
 
+## 1.3.1
+
+- Allow any string for `descriptionSrc`
+
+## 1.3.0
+
+- Added additional default attributes - `className`, `dir`, `exportparts`, `htmlFor`, `hidden`, `id`, `lang`, `ref`, `tabIndex`, `title`, and `translate`.
+
+## 1.2.0
+
+- Add `defaultExport` options to resolve imports without named exports.
+
+## 1.1.1
+
+- Exclude private methods and properties using the `#` access modifier
+
 ## 1.1.0
 
 - Recursively generate output directories