+++ packages/lazy-loader/CHANGELOG.md
@@ -1,5 +1,17 @@
 # CHANGELOG
 
+## 1.2.0
+
+- Add runtime scoping
+
+## 1.1.1
+
+- Update docs
+
+## 1.1.0
+
+- Add eager-loader
+
 ## 1.0.0
 
 - Initial release