+++ packages/react-wrappers/CHANGELOG.md
@@ -1,5 +1,31 @@
 # CHANGELOG
 
+## 1.4.1
+
+- Allow any string for `descriptionSrc`
+
+## 1.4.0
+
+- Added additional default attributes - `classList`, `dir`, `exportparts`, `id`, `lang`, `part`, `title`, and `translate`
+- Fixed attribute mapping in types
+
+## 1.3.0
+
+- Add `defaultExport` options to resolve imports without named exports.
+
+## 1.2.1
+
+- Included file extensions for exports in manifest
+
+## 1.2.0
+
+- Exclude private members using the `#` access modifier
+- Add `htmlFor` global prop
+- Patched issue with `className` prop
+- Patched issue with `style` prop when using style object
+- Removed redundant prop listing
+- Removed attribute utility helpers
+
 ## 1.1.1
 
 - Removed duplicate `ref` prop in wrapper