+++ packages/solidjs-integration/CHANGELOG.md
@@ -1,5 +1,26 @@
 # CHANGELOG
 
+## 1.6.1
+
+- Allow any string for `descriptionSrc`
+
+## 1.6.0
+
+- Added `innerHTML` and `textContent` attributes
+- Removed `htmlFor` attribute as it is not used in SolidJS JSX
+
+## 1.5.0
+
+- Added additional default attributes - `dir`, `exportparts`, `htmlFor`, `hidden`, `id`, `lang`, `ref`, `tabIndex`, `title`, and `translate`.
+
+## 1.4.0
+
+- Add `defaultExport` options to resolve imports without named exports.
+
+## 1.3.1
+
+- Exclude private methods and properties using the `#` access modifier
+
 ## 1.3.0
 
 - Recursively generate output directories