+++ packages/vuejs-integration/CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## 1.1.0
+
+- Add `defaultExport` options to resolve imports without named exports.
+
+## 1.0.1
+
+- Exclude private methods and properties using the `#` access modifier
+
 ## 1.0.0
 
 - Initial release